requirements.txt
pandas
streamlit
streamlit-lottie
numpy
tensorflow
Pillow
